+++ srcs/main.c
@@ -24,20 +24,20 @@ void	enter_shell(char **argv, char **envp)
 	}
 }
 
-int	prompt(t_lexer *lexer)
-{
-	g_state = COMMAND;
-	init_signals(); // ? testing
-	while (lexer->str == NULL || lexer->str[0] == '\0')
-	{
-		lexer->str = readline("momoshell-1.0 🐈 ");
-		if (!lexer->str)
-			exit(EXIT_SUCCESS);
-		if (*lexer->str)
-			add_history(lexer->str);
-	}
-	return (1);
-}
+// int	prompt(t_lexer *lexer)
+// {
+// 	g_state = COMMAND;
+// 	init_signals(); // ? testing
+// 	while (lexer->str == NULL || lexer->str[0] == '\0')
+// 	{
+// 		lexer->str = readline("momoshell-1.0 🐈 ");
+// 		if (!lexer->str)
+// 			exit(EXIT_SUCCESS);
+// 		if (*lexer->str)
+// 			add_history(lexer->str);
+// 	}
+// 	return (1);
+// }
 
 int	main(int argc, char **argv, char **envp)
 {
