+++ srcs/utils/gnl/gnl_utils.c
