--- srcs/utils/libft/ft_isascii.c
@@ -1,16 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_isascii.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/10/23 15:51:20 by nsterk        #+#    #+#                 */
-/*   Updated: 2021/03/03 01:36:23 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-int	ft_isascii(int c)
-{
-	return (c >= 0 && c <= 127);
-}
