--- srcs/utils/libft/ft_isprint.c
@@ -1,16 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_isprint.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <marvin@codam.nl>                     +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/09/20 18:16:05 by nsterk        #+#    #+#                 */
-/*   Updated: 2021/03/03 01:36:35 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-int	ft_isprint(int c)
-{
-	return (c >= 32 && c <= 126);
-}
