--- srcs/utils/libft/ft_lstiter.c
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_lstiter.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/11/03 17:55:29 by nsterk        #+#    #+#                 */
-/*   Updated: 2020/11/05 14:04:19 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_lstiter(t_list *lst, void (*f)(void *))
-{
-	t_list	*temp;
-
-	if (!lst || !f)
-		return ;
-	temp = lst;
-	while (temp != NULL)
-	{
-		f(temp->content);
-		temp = temp->next;
-	}
-}
