--- srcs/utils/libft/ft_lstlast.c
@@ -1,26 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_lstlast.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/11/03 15:29:51 by nsterk        #+#    #+#                 */
-/*   Updated: 2022/09/22 17:38:02 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "utils.h"
-#include "lexer.h"
-
-t_list	*ft_lstlast(t_list *lst)
-{
-	t_list	*current;
-
-	if (!lst)
-		return (NULL);
-	current = lst;
-	while (current->next != NULL)
-		current = current->next;
-	return (current);
-}
