--- srcs/utils/libft/ft_lstnew.c
@@ -1,41 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_lstnew.c                                        :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/11/03 14:04:12 by nsterk        #+#    #+#                 */
-/*   Updated: 2022/09/22 17:38:18 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "utils.h"
-#include "lexer.h"
-
-t_list	*ft_lstnew(void *content)
-{
-	t_list	*new_element;
-
-	new_element = malloc(sizeof(t_list));
-	if (!new_element)
-		return (NULL);
-	new_element->content = content;
-	new_element->next = NULL;
-	return (new_element);
-}
-
-// t_token	*token_new(char *word)
-// {
-// 	t_token	*new;
-
-// 	new = malloc(sizeof(t_token));
-// 	if (!new)
-// 		return (NULL);
-// 	new->prev = NULL;
-// 	new->word = ft_strdup((const char *)word);
-// 	if (!new->word)
-// 		return (NULL);
-// 	new->next = NULL;
-// 	return (new);
-// }
