--- srcs/utils/libft/ft_memset.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_memset.c                                        :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/10/26 20:09:15 by nsterk        #+#    #+#                 */
-/*   Updated: 2020/11/05 14:13:25 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	*ft_memset(void *str, int c, size_t n)
-{
-	size_t			i;
-	unsigned char	*ptr;
-
-	i = 0;
-	ptr = str;
-	while (i < n)
-	{
-		ptr[i] = c;
-		i++;
-	}
-	return (str);
-}
