+++ srcs/utils/libft/ft_strcheck.c
