--- srcs/utils/libft/ft_strclen.c
@@ -1,29 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_strclen.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <marvin@codam.nl>                     +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/09/16 12:26:39 by nsterk        #+#    #+#                 */
-/*   Updated: 2021/03/24 16:17:45 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include <stdlib.h>
-
-size_t	ft_strclen(const char *s, char c)
-{
-	size_t	i;
-
-	i = 0;
-	if (s)
-	{
-		while (s[i] != '\0')
-		{
-			if (s[i] != c)
-				i++;
-		}
-	}
-	return (i);
-}
