--- srcs/utils/libft/ft_strcmp.c
@@ -1,26 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_strcmp.c                                        :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <marvin@codam.nl>                     +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/09/21 17:08:30 by nsterk        #+#    #+#                 */
-/*   Updated: 2021/03/03 01:41:14 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-int	ft_strcmp(char *s1, char *s2)
-{
-	int	i;
-
-	i = 0;
-	while (s1[i] != '\0' || s2[i] != '\0')
-	{
-		if (s1[i] == s2[i])
-			i++;
-		else
-			return (s1[i] - s2[i]);
-	}
-	return (0);
-}
