+++ srcs/utils/libft/ft_strndup.c
@@ -1,16 +1,4 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_strndup.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2022/09/14 15:31:54 by nsterk        #+#    #+#                 */
-/*   Updated: 2022/09/14 15:32:18 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
+#include "utils.h"
 
 char	*ft_strndup(char *src, size_t n)
 {
