--- srcs/utils/libft/ft_strnlen.c
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_strnlen.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/11/12 21:16:10 by nsterk        #+#    #+#                 */
-/*   Updated: 2020/11/12 21:36:05 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include <stdlib.h>
-
-size_t	ft_strnlen(const char *s, size_t maxlen)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < maxlen)
-	{
-		if (s[i] == '\0')
-			return (i);
-		i++;
-	}
-	return (maxlen);
-}
