--- srcs/utils/libft/ft_tolower.c
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_tolower.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <marvin@codam.nl>                     +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/09/20 21:13:09 by nsterk        #+#    #+#                 */
-/*   Updated: 2020/10/27 10:03:21 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-int	ft_tolower(int c)
-{
-	if (c >= 'A' && c <= 'Z')
-		c += 32;
-	return (c);
-}
