--- srcs/utils/libft/ft_toupper.c
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   ft_toupper.c                                       :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: nsterk <marvin@codam.nl>                     +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2020/09/20 18:18:41 by nsterk        #+#    #+#                 */
-/*   Updated: 2020/10/27 10:04:09 by nsterk        ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-int	ft_toupper(int c)
-{
-	if (c >= 'a' && c <= 'z')
-		c -= 32;
-	return (c);
-}
